rename: view render file name
