+++ src/features/renderNews/components/list/index.js
