+++ src/features/renderNews/utils/renderView.js
@@ -1,6 +1,6 @@
 import { MainNewsState } from "../../../types/news.js";
-import { renderGridView } from "../components/grid/renderNewsGrid.js";
-import { renderListView } from "../components/list/renderNewsList.js";
+import { renderGridView } from "../components/grid/index.js";
+import { renderListView } from "../components/list/index.js";
 
 const renderView = {
   ["list-view"]: renderListView,
